+++ src/hooks/index.ts
@@ -7,3 +7,4 @@ export * from "./useAddPacelineRides";
 export * from "./useEditRide";
 export * from "./useDeleteRide";
 export * from "./useUpdateUser";
+export * from "./useLocalStorage";