+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.14.0]
+
+### Added
+
+- Store anonymous rider details in their browser (localStorage)
+- Use this to show whether they are going to rides and to auto-fill join form
+
 ## [0.13.0]
 
 ### Added